[Silver III] Title: 주유소, Time: 20 ms, Memory: 2800 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -1,6 +1,6 @@
 #include <bits/stdc++.h>
 using namespace std;
-int N, dist[100'005], fee, minFee = 1'000'000'005, res;
+long long N, dist[100'005], fee, minFee = 1'000'000'005, res;
 int main() 
 {
     ios::sync_with_stdio(false);
